# docs/swagger.yaml
definitions:
  diploma_modules_auth_handler_model.LoginInput:
    properties:
      password:
        example: secure123
        type: string
      phone_number:
        example: "+123456789"
        type: string
    type: object
  diploma_modules_auth_handler_model.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  diploma_modules_auth_handler_model.RegisterInput:
    properties:
      confirm_password:
        example: secure123
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: secure123
        type: string
      phone_number:
        example: "+123456789"
        type: string
    type: object
  diploma_modules_auth_handler_model.RegisterResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  diploma_modules_cart_handler_model.AddProductToCartInput:
    properties:
      customer_id:
        type: integer
      price:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      supplier_id:
        type: integer
    type: object
  diploma_modules_cart_handler_model.ErrorResponse:
    properties:
      error: {}
    type: object
  diploma_modules_cart_handler_model.GetCartResponse:
    properties:
      customer_id:
        type: integer
      suppliers:
        items:
          $ref: '#/definitions/model.Supplier'
        type: array
      total:
        type: integer
    type: object
  diploma_modules_product_handler_model.DetailedProduct:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
      suppliers:
        items:
          $ref: '#/definitions/diploma_modules_product_handler_model.ProductSupplierInfo'
        type: array
    type: object
  diploma_modules_product_handler_model.ProductListResponse:
    properties:
      product_list:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      total:
        type: integer
    type: object
  diploma_modules_product_handler_model.ProductResponse:
    properties:
      product:
        $ref: '#/definitions/diploma_modules_product_handler_model.DetailedProduct'
    type: object
  diploma_modules_product_handler_model.ProductSupplierInfo:
    properties:
      delivery_fee:
        type: number
      minimum_free_delivery_amount:
        type: number
      name:
        type: string
      supplier_id:
        type: integer
    type: object
  gin.H:
    additionalProperties: {}
    type: object
  model.Product:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
    type: object
  model.Supplier:
    properties:
      delivery_fee:
        type: integer
      free_delivery_amount:
        type: integer
      id:
        type: integer
      minOrderAmount:
        type: integer
      name:
        type: string
      product_list:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      total_amount:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login user and return tokens
      parameters:
      - description: Login input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/diploma_modules_auth_handler_model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diploma_modules_auth_handler_model.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      summary: User login
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/diploma_modules_auth_handler_model.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/diploma_modules_auth_handler_model.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: User registration
      tags:
      - auth
  /api/card/put:
    get:
      consumes:
      - application/json
      description: --
      parameters:
      - description: Put Card input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/diploma_modules_cart_handler_model.AddProductToCartInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
      summary: Put product to Card
      tags:
      - cart
  /api/product/:id:
    get:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: product id
        in: query
        name: product_id
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/diploma_modules_product_handler_model.ProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: User registration
      tags:
      - product
  /api/product/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of products with pagination support using limit
        and offset
      parameters:
      - description: Limit number of products
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diploma_modules_product_handler_model.ProductListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get product list
      tags:
      - product
    post:
      consumes:
      - application/json
      description: get customer's cart
      parameters:
      - description: customer ID
        in: query
        name: customer-ID
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.GetCartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_cart_handler_model.ErrorResponse'
      summary: Get cart
      tags:
      - cart
swagger: "2.0"